feat(transfer): update transfer table
