+++ crates/ark-transfers/src/transfer.rs
@@ -1,57 +1,13 @@
 use super::mint::{process_mint_event, TokenData, TransactionData};
-use ark_db::token::update::update_token;
+use super::utils::get_token_uri;
+use ark_db::token_event::create::{create_token_event, TokenEvent};
 use ark_starknet::utils::TokenId;
-use ark_starknet::{
-    client::get_block_with_txs, client::get_token_owner, utils::get_contract_property_string,
-};
+use ark_starknet::{client::get_block_with_txs, client::get_token_owner};
 use aws_sdk_dynamodb::Client as DynamoClient;
 use log::info;
 use reqwest::Client as ReqwestClient;
 use starknet::core::types::{EmittedEvent, FieldElement};
 use std::error::Error;
-
-async fn get_token_uri(
-    client: &ReqwestClient,
-    token_id_low: u128,
-    token_id_high: u128,
-    contract_address: &str,
-    block_number: u64,
-) -> String {
-    info!("get_token_id: [{:?}, {:?}]", token_id_low, token_id_high);
-
-    let token_id_low_hex = format!("{:x}", token_id_low);
-    let token_id_high_hex = format!("{:x}", token_id_high);
-
-    let token_uri_cairo0 = get_contract_property_string(
-        client,
-        contract_address,
-        "tokenURI",
-        vec![&token_id_low_hex, &token_id_high_hex],
-        block_number,
-    )
-    .await;
-
-    if token_uri_cairo0 != "undefined" && !token_uri_cairo0.is_empty() {
-        return token_uri_cairo0;
-    }
-
-    let token_uri = get_contract_property_string(
-        client,
-        contract_address,
-        "token_uri",
-        vec![&token_id_low_hex, &token_id_high_hex],
-        block_number,
-    )
-    .await;
-
-    info!("token_uri: {:?}", token_uri);
-
-    if token_uri != "undefined" && !token_uri.is_empty() {
-        return token_uri;
-    }
-
-    "undefined".to_string()
-}
 
 pub async fn process_transfers(
     client: &ReqwestClient,
@@ -71,7 +27,6 @@ pub async fn process_transfers(
     let timestamp = block.get("timestamp").unwrap().as_u64().unwrap();
 
     // Extracting "data" from event
-
     let from_address_field_element = event.data[0];
     let from_address = format!("{:#064x}", from_address_field_element);
     let to_address = format!("{:#064x}", event.data[1]);
@@ -145,16 +100,19 @@ pub async fn process_transfers(
         )
         .await;
     } else {
-        update_token(
-            dynamo_db_client,
-            contract_address.to_string(),
-            formated_token_id.padded_token_id.as_str(),
-            from_address.to_string(),
-            to_address.to_string(),
-            &timestamp.clone(),
-            transaction_hash.to_string(),
-        )
-        .await?;
+        let token_event = TokenEvent {
+            address: contract_address.clone(),
+            timestamp,
+            block_number,
+            event_type: "transfer".to_string(),
+            from_address: from_address.to_string(),
+            padded_token_id: formated_token_id.padded_token_id.clone(),
+            token_uri: token_uri.clone(),
+            to_address: to_address.to_string(),
+            transaction_hash: transaction_hash.clone(),
+            token_type: contract_type.to_string(),
+        };
+        let _ = create_token_event(dynamo_db_client, token_event).await;
     }
 
     Ok(())
